Rearrange velocity
