+++ src/Phase/quadrature.jl
@@ -5,6 +5,57 @@
 """
 $(SIGNATURES)
 
+Maxwell quadrature
+
+# Arguments
+* ``N``: quadrature order (MUST less than 33)
+"""
+function maxwell_quadrature(N::Integer, C = 1::Real)
+    @assert N <= 33
+
+    py"""
+    import numpy as np
+    from numpy import linalg as LA
+
+    def dvGH(N2,C):
+        N = N2//2
+
+        a = np.zeros(N)
+        b = np.zeros(N)
+        a[0] = 1.0/np.sqrt(np.pi)
+        a[1] = 2.0/np.sqrt(np.pi)/(np.pi-2.0)
+        b[1] = a[0]/( a[0] + a[1])/2.0
+
+        for i in range(2,N):
+            b[i] = (i-1)+1.0/2.0-b[i-1]-a[i-1]**2
+            a[i] = (i**2/4.0/b[i]-b[i-1]-1.0/2)/a[i-1]-a[i-1]
+
+        J = np.diag(a) + np.diag(np.sqrt(b[1:N]),1) \
+        + np.diag(np.sqrt(b[1:N]),-1)
+
+        v,V = LA.eig(J)
+
+        w = V[0,:]*V[0,:]*np.sqrt(np.pi)/2.0
+
+        vw = np.transpose(np.vstack((v,w)))
+        vw = vw[vw[:,0].argsort()]
+        v = vw[:,0]
+        w = vw[:,1]
+
+        Xis = np.hstack((-np.flipud(v),v))
+        weights = np.hstack((np.flipud(w),w))
+        weights = weights*np.exp(Xis**2)*C
+        Xis = Xis*C
+        return (Xis, weights)
+    """
+
+    p, w = py"dvGH"(N, C)
+end
+
+
+"""
+$(SIGNATURES)
+
 Gauss-Legendre quadrature
 
 # Arguments
@@ -14,8 +65,7 @@ Gauss-Legendre quadrature
 * ``points``: quadrature points in 3D coordinate
 * ``weights``: quadrature weights
 """
-function legendre_quadrature(n::T) where {T<:Integer}
-
+function legendre_quadrature(n::Integer)
     pointsxyz = zeros(n * n, 3)
     weights = zeros(n * n)
 
@@ -38,7 +88,6 @@ function legendre_quadrature(n::T) where {T<:Integer}
     weights = weights[:]
 
     return pointsxyz, weights
-
 end
 
 
@@ -54,8 +103,14 @@ Octaeder quadrature
 # Outputs
 * points and triangulation
 """
-function octa_quadrature(n::T, slerpflag = true::Bool) where {T<:Integer}
+function octa_quadrature(n::Integer, slerpflag = true::Bool)
+    points, triangulation = octa_triangle(n, slerpflag)
+    weights = triangle_weights(points, triangulation)
 
+    return points, weights
+end
+
+function octa_triangle(n::Integer, slerpflag = true::Bool)
     # integral range
     pt0 = [0.0, 0.0, 1.0]
     pt1 = [0.0, 1.0, 0.0]
@@ -162,7 +217,6 @@ function octa_quadrature(n::T, slerpflag = true::Bool) where {T<:Integer}
     triangulation = permutedims(triangulation)
 
     return points, triangulation
-
 end
 
 
@@ -178,7 +232,7 @@ Create quadrature weights from points and triangulation
 # Outputs
 * quadrature weights
 """
-function quadrature_weights(xyz::X, triangles::Y) where {X<:AA{<:Real,2},Y<:AA{<:Integer,2}}
+function triangle_weights(xyz::AM, triangles::AM)
     weights = zeros(axes(xyz, 1))
     nTriangles = size(triangles, 1)
     xy = zeros(3)
@@ -225,4 +279,24 @@ function quadrature_weights(xyz::X, triangles::Y) where {X<:AA{<:Real,2},Y<:AA{<
     end
 
     return weights
+end
+
+
+"""
+$(SIGNATURES)
+
+Evaluate quadrature weight from Newton-Cotes rule
+"""
+function newton_cotes(idx::T, num::T) where {T<:Integer}
+    if idx == 1 || idx == num
+        nc_coeff = 14.0 / 45.0
+    elseif (idx - 5) % 4 == 0
+        nc_coeff = 28.0 / 45.0
+    elseif (idx - 3) % 4 == 0
+        nc_coeff = 24.0 / 45.0
+    else
+        nc_coeff = 64.0 / 45.0
+    end
+
+    return nc_coeff
 end
