+++ src/Phase/velocity.jl
@@ -607,26 +607,6 @@ struct UnstructVSpace{TR<:Union{Real,AV},TI<:Union{Integer,AV},TA<:AA,TB<:AV{<:R
     weights::TB
 end
 
-
-"""
-$(SIGNATURES)
-
-Evaluate quadrature weight from Newton-Cotes rule
-"""
-function newton_cotes(idx::T, num::T) where {T<:Integer}
-    if idx == 1 || idx == num
-        nc_coeff = 14.0 / 45.0
-    elseif (idx - 5) % 4 == 0
-        nc_coeff = 28.0 / 45.0
-    elseif (idx - 3) % 4 == 0
-        nc_coeff = 24.0 / 45.0
-    else
-        nc_coeff = 64.0 / 45.0
-    end
-
-    return nc_coeff
-end
-
 # ------------------------------------------------------------
 # Extended Base.show()
 # ------------------------------------------------------------
@@ -656,47 +636,4 @@ function Base.show(io::IO, vs::VSpace3D{TR,TI,TA}) where {TR,TI,TA}
         "domain: ($(vs.u0),$(vs.u1)) × ($(vs.v0),$(vs.v1)) × ($(vs.w0),$(vs.w1))\n",
         "resolution: $(vs.nu) × $(vs.nv) × $(vs.nw)\n",
     )
-end
-
-
-function maxwell_quadrature(N::Integer, C = 1::Real)
-    @assert N <= 33
-
-    py"""
-    import numpy as np
-    from numpy import linalg as LA
-
-    def dvGH(N2,C):
-        N = N2//2
-
-        a = np.zeros(N)
-        b = np.zeros(N)
-        a[0] = 1.0/np.sqrt(np.pi)
-        a[1] = 2.0/np.sqrt(np.pi)/(np.pi-2.0)
-        b[1] = a[0]/( a[0] + a[1])/2.0
-
-        for i in range(2,N):
-            b[i] = (i-1)+1.0/2.0-b[i-1]-a[i-1]**2
-            a[i] = (i**2/4.0/b[i]-b[i-1]-1.0/2)/a[i-1]-a[i-1]
-
-        J = np.diag(a) + np.diag(np.sqrt(b[1:N]),1) \
-        + np.diag(np.sqrt(b[1:N]),-1)
-
-        v,V = LA.eig(J)
-
-        w = V[0,:]*V[0,:]*np.sqrt(np.pi)/2.0
-
-        vw = np.transpose(np.vstack((v,w)))
-        vw = vw[vw[:,0].argsort()]
-        v = vw[:,0]
-        w = vw[:,1]
-
-        Xis = np.hstack((-np.flipud(v),v))
-        weights = np.hstack((np.flipud(w),w))
-        weights = weights*np.exp(Xis**2)*C
-        Xis = Xis*C
-        return (Xis, weights)
-    """
-
-    p, w = py"dvGH"(N, C)
 end
